Init casper-consensus, casper-primitives crate
