+++ substrate/casper/consensus/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "substrate-casper"
+version = "0.1.0"
+authors = ["Wei Tang <hi@that.world>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
