+++ substrate/casper/srml/Cargo.toml
@@ -9,8 +9,9 @@ serde = { version = "1.0", optional = true }
 codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false }
 srml-support = { git = "https://github.com/paritytech/substrate", default-features = false }
 system = { package = "srml-system", git = "https://github.com/paritytech/substrate", default-features = false }
-balances = { package = "srml-balances", git = "https://github.com/paritytech/substrate", default-features = false }
+session = { package = "srml-session", git = "https://github.com/paritytech/substrate", default-features = false }
 sr-primitives = { git = "https://github.com/paritytech/substrate", default-features = false }
+casper-primitives = { package = "substrate-casper-primitives", path = "../primitives", default-features = false }
 
 [dev-dependencies]
 sr-io = { git = "https://github.com/paritytech/substrate" }
@@ -22,7 +23,8 @@ std = [
 	"serde",
 	"codec/std",
 	"sr-primitives/std",
+	"casper-primitives/std",
 	"srml-support/std",
 	"system/std",
-	"balances/std",
+	"session/std",
 ]
