+++ substrate/casper/srml/src/lib.rs
@@ -1,0 +1,198 @@
+// Copyright 2017-2019 Parity Technologies (UK) Ltd.
+// This file is part of Substrate.
+
+// Substrate is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Substrate is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Substrate.  If not, see <http://www.gnu.org/licenses/>.
+
+//! # Casper Consensus Module
+
+#![cfg_attr(not(feature = "std"), no_std)]
+
+use srml_support::{StorageValue, dispatch::Result, decl_module, decl_storage, decl_event};
+use system::{ensure_signed, ensure_root};
+use sr_primitives::weights::SimpleDispatchInfo;
+use casper_primitives::{ValidatorId, ValidatorWeight};
+
+/// Casper module's configuration trait.
+pub trait Trait: system::Trait {
+	/// The overarching event type.
+	type Event: From<Event> + Into<<Self as system::Trait>::Event>;
+}
+
+decl_storage! {
+	trait Store for Module<T: Trait> as Casper {
+		Validators get(validators) config(): Vec<(ValidatorId, ValidatorWeight)>;
+	}
+}
+
+decl_event!(
+	/// Casper events.
+	pub enum Event {
+		// Just a normal `enum`, here's a dummy event to ensure it compiles.
+		/// Dummy event, just here so there's a generic type that's used.
+		Dummy,
+	}
+);
+
+decl_module! {
+	pub struct Module<T: Trait> for enum Call where origin: T::Origin {
+		fn deposit_event() = default;
+
+		// The signature could also look like: `fn on_initialize()`
+		fn on_initialize(_n: T::BlockNumber) {
+			// Anything that needs to be done at the start of the block.
+			// We don't do anything here.
+		}
+
+		// The signature could also look like: `fn on_finalize()`
+		fn on_finalize(_n: T::BlockNumber) {
+			// Anything that needs to be done at the end of the block.
+			// We just kill our dummy storage item.
+		}
+
+		// A runtime code run after every block and have access to extended set of APIs.
+		//
+		// For instance you can generate extrinsics for the upcoming produced block.
+		fn offchain_worker(_n: T::BlockNumber) {
+			// We don't do anything here.
+			// but we could dispatch extrinsic (transaction/unsigned/inherent) using
+			// runtime_io::submit_extrinsic
+		}
+	}
+}
+
+// The main implementation block for the module. Functions here fall into three broad
+// categories:
+// - Public interface. These are functions that are `pub` and generally fall into inspector
+// functions that do not write to storage and operation functions that do.
+// - Private functions. These are your usual private utilities unavailable to other modules.
+impl<T: Trait> Module<T> {
+}
+
+#[cfg(test)]
+mod tests {
+	use super::*;
+
+	use srml_support::{assert_ok, impl_outer_origin, parameter_types};
+	use sr_io::with_externalities;
+	use primitives::{H256, Blake2Hasher};
+	// The testing primitives are very useful for avoiding having to work with signatures
+	// or public keys. `u64` is used as the `AccountId` and no `Signature`s are requried.
+	use sr_primitives::{
+		Perbill, traits::{BlakeTwo256, OnInitialize, OnFinalize, IdentityLookup}, testing::Header
+	};
+
+	impl_outer_origin! {
+		pub enum Origin for Test {}
+	}
+
+	// For testing the module, we construct most of a mock runtime. This means
+	// first constructing a configuration type (`Test`) which `impl`s each of the
+	// configuration traits of modules we want to use.
+	#[derive(Clone, Eq, PartialEq)]
+	pub struct Test;
+	parameter_types! {
+		pub const BlockHashCount: u64 = 250;
+		pub const MaximumBlockWeight: u32 = 1024;
+		pub const MaximumBlockLength: u32 = 2 * 1024;
+		pub const AvailableBlockRatio: Perbill = Perbill::one();
+	}
+	impl system::Trait for Test {
+		type Origin = Origin;
+		type Index = u64;
+		type BlockNumber = u64;
+		type Hash = H256;
+		type Call = ();
+		type Hashing = BlakeTwo256;
+		type AccountId = u64;
+		type Lookup = IdentityLookup<Self::AccountId>;
+		type Header = Header;
+		type WeightMultiplierUpdate = ();
+		type Event = ();
+		type BlockHashCount = BlockHashCount;
+		type MaximumBlockWeight = MaximumBlockWeight;
+		type MaximumBlockLength = MaximumBlockLength;
+		type AvailableBlockRatio = AvailableBlockRatio;
+	}
+	parameter_types! {
+		pub const ExistentialDeposit: u64 = 0;
+		pub const TransferFee: u64 = 0;
+		pub const CreationFee: u64 = 0;
+		pub const TransactionBaseFee: u64 = 0;
+		pub const TransactionByteFee: u64 = 0;
+	}
+	impl balances::Trait for Test {
+		type Balance = u64;
+		type OnFreeBalanceZero = ();
+		type OnNewAccount = ();
+		type Event = ();
+		type TransactionPayment = ();
+		type TransferPayment = ();
+		type DustRemoval = ();
+		type ExistentialDeposit = ExistentialDeposit;
+		type TransferFee = TransferFee;
+		type CreationFee = CreationFee;
+		type TransactionBaseFee = TransactionBaseFee;
+		type TransactionByteFee = TransactionByteFee;
+		type WeightToFee = ();
+	}
+	impl Trait for Test {
+		type Event = ();
+	}
+	type Example = Module<Test>;
+
+	// This function basically just builds a genesis storage key/value store according to
+	// our desired mockup.
+	fn new_test_ext() -> sr_io::TestExternalities<Blake2Hasher> {
+		let mut t = system::GenesisConfig::default().build_storage::<Test>().unwrap();
+		// We use default for brevity, but you can configure as desired if needed.
+		balances::GenesisConfig::<Test>::default().assimilate_storage(&mut t).unwrap();
+		GenesisConfig::<Test>{
+			dummy: 42,
+			// we configure the map with (key, value) pairs.
+			bar: vec![(1, 2), (2, 3)],
+			foo: 24,
+		}.assimilate_storage(&mut t).unwrap();
+		t.into()
+	}
+
+	#[test]
+	fn it_works_for_optional_value() {
+		with_externalities(&mut new_test_ext(), || {
+			// Check that GenesisBuilder works properly.
+			assert_eq!(Example::dummy(), Some(42));
+
+			// Check that accumulate works when we have Some value in Dummy already.
+			assert_ok!(Example::accumulate_dummy(Origin::signed(1), 27));
+			assert_eq!(Example::dummy(), Some(69));
+
+			// Check that finalizing the block removes Dummy from storage.
+			<Example as OnFinalize<u64>>::on_finalize(1);
+			assert_eq!(Example::dummy(), None);
+
+			// Check that accumulate works when we Dummy has None in it.
+			<Example as OnInitialize<u64>>::on_initialize(2);
+			assert_ok!(Example::accumulate_dummy(Origin::signed(1), 42));
+			assert_eq!(Example::dummy(), Some(42));
+		});
+	}
+
+	#[test]
+	fn it_works_for_default_value() {
+		with_externalities(&mut new_test_ext(), || {
+			assert_eq!(Example::foo(), 24);
+			assert_ok!(Example::accumulate_foo(Origin::signed(1), 1));
+			assert_eq!(Example::foo(), 25);
+		});
+	}
+}
